requirements.txt
Django==1.9.8
djangorestframework==3.4.1
MySQL-python==1.2.5
python-memcached==1.58
uwsgi==2.0.14
pytz==2016.6.1
baidu-aip==1.3.2.1
bosonnlp==0.9.0